pyproject.toml
[tool.poetry]
name = "idpy"
version = "0.1.0"
description = "Mock SAML IDP in Python"
authors = ["Cameron Clark <cameron@gremlin.com>"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.1"
pysaml2 = "^7.0.1"
Flask-RESTful = "^0.3.9"
Flask-SQLAlchemy = "^2.5.1"
flask-marshmallow = "^0.14.0"
marshmallow-sqlalchemy = "^0.26.1"

[tool.poetry.dev-dependencies]
black = "^21.6b0"
flake8 = "^3.9.2"
isort = "^5.8.0"
pytest = "^6.2.4"
pytest-flask = "^1.2.0"
mypy = "^0.902"
pytest-cov = "^2.12.1"
pre-commit = "^2.13.0"
beautifulsoup4 = "^4.9.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
